Create abstract base class for rpc integration tests
